requirements.txt
asgiref==3.5.2
django==4.1.3
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
pgcli
Pillow
datetime
pathlib
djangorestframework-simplejwt
boto3
gunicorn
django-storages